--- a/src/backend/metal/src/lib.rs
+++ b/src/backend/metal/src/lib.rs
@@ -10,6 +10,7 @@ extern crate core_graphics;
 extern crate block;
 extern crate smallvec;
 extern crate spirv_cross;
+extern crate storage_map;
 
 #[cfg(feature = "antidote")]
 extern crate antidote as lock;