Fixes #2227
Adds dependency to `storage-map` crate for locking on Metal. If it proves to be useful, we'll likely want it to be used for other backends as well.

Dota performance changes:
  - immediate: 90 -> 92 fps
  - deferred: 92 -> 99 fps

PR checklist:
- [ ] `make` succeeds (on *nix)
- [x] `make reftests` succeeds
- [ ] tested examples with the following backends:
- [ ] `rustfmt` run on changed code